+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - `--analytic_marg` option in `pyvela` script
 - `--analytic_marg all` option in `pyvela` script
 - Add FD and FDJUMP to analytically marginalizable parameters
+- Script to plot chains
 - `SPNTA.maxpost_params` property
 - Pre-commit hooks
 ## Changed